randVector
